*FIXED* Crash on login due to messaging system. This really should be an LL_ERRS....
--HG--
extra : convert_revision : svn%3A1bd1a86a-1eb0-4ff1-abcc-96d0af321e0c/trunk%401498

+++ indra/llmessage/llmessagetemplate.h
@@ -125,7 +125,21 @@ public:
 	}
 	~LLMsgData()
 	{
-		for_each(mMemberBlocks.begin(), mMemberBlocks.end(), DeletePairedPointer());
+		if(mMemberBlocks.end()->first)
+		{
+			for_each(mMemberBlocks.begin(), mMemberBlocks.end(), DeletePairedPointer());
+		}
+		else
+		{
+			if(mName)
+			{
+				llwarns << "Bad blocks in MsgData named " << mName << llendl;
+			}
+			else
+			{
+				llwarns << "Bad blocks in MsgData of unknown name?" << llendl;
+			}
+		}
 	}
 
 	void addBlock(LLMsgBlkData *blockp)
